Adding crypt for user/password_digest
Setting validation for minimums for password length, verification.
Used db migration to add a new column to user models database table, of
“password_digest”, which is string data type.

NEED TO: Go back and review password_confirmation implementation, as it
is using a gem.  It seems like at of first glance, password is another
string entry in our User model, and there should be a validation that
is MISSING, which should be that password == password_confirmation.

Unless this is already built into password_digest.