Stringing them together, the rails console work
Using the rails console of User.find(1) to def each of the shows.

In rails console, we used “User.find()” or “User.find_by()” to find who
we were looking for. Therefore, in our show controller, we find out
which @user (isntance variable) that we’re looking for, by using the
same ID as it was when created. So user 1 is the first user I made and
found in console by User.find(1).

Therefore, in the show controller, I have to tell RoR which @user I’m
looking for.

Define @user = User.find(params[:id]).

The id is integer 1-X, depending on which one was created and when.